# Authentication/Interfaces/REST/Resources/LoginResource.cs
namespace HealMeAppBackend.API.Authentication.Interfaces.REST.Resources
{
    public class LoginResource
    {
        public string Username { get; set; }
        public string Password { get; set; }
    }
}