# Authentication/Interfaces/REST/Resources/RegisterResource.cs
namespace HealMeAppBackend.API.Authentication.Interfaces.REST.Resources
{
    public class RegisterResource
    {
        public string Username { get; set; }
        public string Password { get; set; }
    }
}